+++ src/icons/question-ghost-icon.tsx
@@ -1,0 +1,24 @@
+/*
+ * Copyright (c) 2011-2021 Genestack Limited
+ * All Rights Reserved
+ * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF GENESTACK LIMITED
+ * The copyright notice above does not evidence any
+ * actual or intended publication of such source code.
+ */
+import * as React from 'react';
+
+import {createIcon} from '../components/icon';
+
+/**
+ * Question Ghost Icon
+ */
+export const QuestionGhostIcon = createIcon(
+    <svg width="20" height="20" viewBox="0 0 20 20" fill="none">
+        <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M16 10a6 6 0 11-12 0 6 6 0 0112 0zm1 0a7 7 0 11-14 0 7 7 0 0114 0zM9.08 7.27c-.19.146-.33.364-.33.73a.75.75 0 01-1.5 0c0-.837.36-1.494.92-1.922C8.703 5.67 9.374 5.5 10 5.5c.638 0 1.308.188 1.837.6.547.428.913 1.08.913 1.9 0 1.13-.78 1.712-1.25 2.062l-.05.038c-.538.404-.7.57-.7.9h-1.5c0-1.13.78-1.712 1.25-2.062l.05-.038c.538-.404.7-.57.7-.9 0-.335-.134-.56-.337-.718C10.692 7.11 10.363 7 10 7c-.375 0-.704.104-.92.27zM11 13a1 1 0 11-2 0 1 1 0 012 0z"
+            className="gs-svg-icon-fill"
+        />
+    </svg>
+);